pyproject.toml
[project]
name = "portfolio-system"
version = "0.1.0"
description = "Portfolio management system with FastAPI backend and React frontend"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn==0.30.1",
    "pydantic==2.8.2",
    "pandas==2.2.2",
    "PyYAML==6.0.2",
    "tushare>=1.2.89",
    "pytest==8.3.2",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]